clubhouse: Refactor ClubhouseView to overlay several fixed widgets
In a future commit, a custom tooltip is needed to be added on top
of a QuestSetButton at a fixed position.

A new custom widget called FixedLayerGroup which allows to overlay
several layers. Each layer will be always a Gtk.Fixed.

To support overlaying several fixed widgets, the old ClubhouseView
is renamed to ClubhouseViewMainLayer and the ClubhouseView is
refactored to inherit from FixedLayerGroup having as the first
layer the old ClubhouseView.

https://phabricator.endlessm.com/T27843