clubhouse: Hide achievements view when clicking outside
There was a margin around the _user_box widget that was preventing
events from being received on the background. Because of that when
clicking into the margin area around that pixel, the achievements
view did not autohide.

This commit also removes that margin and fixes that issue by adding
a new event box wrapping the box inside the _user_box.

https://phabricator.endlessm.com/T27995