Fix mapping characters to quests
The quests were defining the character instead of the pathway. This
change does it the other way round, and simplifies the subclassing of
QuestSet, leaving the pathways as the only QuestSet, and merging the
character mission into them.

This updates all the tags in quests. Also adjusts the import script
and the tests.

https://phabricator.endlessm.com/T28151