system: Don't override hack bg setting on first disable
There's a race condition with the first contact quest. The hack mode is
enabled and that emits a signal that changes the background.

But the first contact quest sets the background to the default one until
the quest is finished, so if the call from the quest is done before the
signal callback, the background setting is set to the current background
and when this method is called again, the background will not change
because the hack setting has the same background that we've right now.

This patch adds a new param to the set_hack_mode to block the signal
handler to be called. It's also needed to call to the set_hack_mode when
the first quest finish to ensure that all the signals are emited at the
end and we've the clubhouse ON with the correct animations.

https://phabricator.endlessm.com/T28046