profile: Prevent unneeded scrollbar to appear in AchievementsView
Before this commit, if a badge had a long description, the
SUMMARY page increased its height, making the scrollbar to appear.
After that, if you switched back the GRID view, the scrollbar
was kept even when there was only one badge in the view.

This commit, prevents that GRID view scrollbar to appear when it
shouldn't.

Note: Each stack page has its own GtkScrolledWindow with each
respective viewport. If you decide to modify certain properties in
the GtkScrolledWindow or in the viewport, you may want to apply
the same properties to the both scrolled windows.

https://phabricator.endlessm.com/T28138