quests: Disable the clubhouse while offering the quickstart
And make sure to enable it again right after the player either accepts
or declines the quickstart. Also enable the app again if the quest is
aborted, for instance if the shell dialogue is closed using the x
button.

Also, auto-offer this quest right away. It was offered after 3 seconds
before. But in those 3 seconds the player could press the hack mode
switcher.

https://phabricator.endlessm.com/T28087