quests, hack2, T2: start app directly and then instruct
When a user clicks on the card and the quest is laucnhed
the app is started right away and then the instructions
follow.

This will help the user making that context switch to the
new app and also prevents from badges on the cards and
achievement badges being given right away as those are
granted and updated at the end of the quest.

https://phabricator.endlessm.com/T28379