clubhouse: Force hack background in hack activation
This patch removes the hack-mode-enabled-picture-uri setting and set the
hack background always the hack mode is enabled.

The custom user background will be stored and set back when the hack
mode is disabled.

If the user changes the background during hack mode and then disables
the hack mode, the custom user background will be keep. But if user
enables the hack mode again, the hack background will be set.

https://phabricator.endlessm.com/T28472