clubhouse, utils: Decouple Hack mode from hack switcher
In the clubhouse everything was binded to the `hack-mode-enabled`
setting. Now we want to let the players try the switcher in the middle
of the Quickstart quest. So we need to do a special case for this
quest. This requires decoupling the switcher and all UI effects from
the `hack-mode-enabled` setting.

- A new `lights-on` property is added to the clubhouse state for
  this. All visual effects are binded to this property.

- When the `hack-mode-enabled` setting changes, `lights-on` will also
  be syncronized.

- The switcher sets the `hack-mode-enabled` setting like before,
  except if the Quickstart quest is running, in which case it only
  applies the UI effects by setting the `lights-on` property.

https://phabricator.endlessm.com/T28087