quests: Use is_app_in_foreground in migration
The wait_for_app_in_foreground is causing some problems in the migration
quest, when using it the window open takes more than 5 seconds to
appear.

It's better to use a while loop with the is_app_in_foreground check and
a pause meantime we discover what's causing the big delay.

https://phabricator.endlessm.com/T28521