data: Resize badges svg files and square the doc
The viewbox for all the svg files is too small and the shell doesn't
render the achievements notification correctly.

This patch resized the svg to have 200x200 document size, so it's bigger
enougth and it's square. The aspect ratio of all badges is the same, the
only thing that's square is the svg document. The extra espace is
transparent.

https://phabricator.endlessm.com/T28505