clubhouse: Prevent from stopping the migration quest with switch
During a migration quest, Using the hack switch will not stop the
migration quest but just set the shell dialogues to invisible.

If you are in a migration quest and set the hack switch off, you
will see that the shell dialogue dissapears. That does not mean
that the quest stopped, just that the dialogue has been set to
non-visible, which is the default behavior of the GNOME Shell.

The GNOME Shell, by default, sets the dialogues to non-visible
when hack-mode-enabled is set to false. So, if you press the
switch button again, you will see exactly the last message before
switching it off.

https://phabricator.endlessm.com/T28473