clubhouse: unlock System app for Hack 1 users during migration
The old System app has been replaced under the
hack_computer namespace. We agreed on granting the Hack 1
users all of the hacking affordences for Hack 1 toy apps.

We only want to grant this for Hack 1 users. Eben though the
"system tour" quest currently unlocks the toolbox
during the quest we might want to cange this in the
near future.

https://phabricator.endlessm.com/T28554