clubhouse: Add QuestRunner class
The ClubhouseView has been a bloater class for so long. This UI widget
contained everything related to quest handling, making it impossible
to use the quest system without a window instance. Even in the
FirstQuest, which doesn't require the Clubhouse open, the window was
instantiated.

This refactor moves everything related to quest handling to a new
QuestRunner class. Widgets can call QuestRunner methods or connect to
its signals through the main application, like this:

    app = Gio.Application.get_default()
    current_quest = app.quest_runner.running_quest
    app.quest_runner.connect('notify::current-episode',
                             self._current_episode_changed_cb)

Changes included in this refactor:

- The app and all widgets handle quests through the app.quest_runner
  instance.
- The style is initialized when the window is ensured.
- There is no special case for FirstQuest.
- FirstQuest displaying the clubhouse window when finished, happens
  now in the QuestRunner.
- The workaround for T28479 is removed.

Changes not included (to prevent an even bigger refactor):

- The MessageBox is still accessed from the QuestRunner to display
  narrative quests. This should be refactored too.

- The scale is being stored in every widget. It should be a property
  of the window and passed as reference to its children.

https://phabricator.endlessm.com/T28479