utils: Don't insert hyphens in some shell dialogue's text
Specifically, in URLs and inline code. For this, these texts are now
wrapped with a <span insert_hyphens="false"> to instruct Pango to
avoid hyphenation on them.

Also: Add back underline and color to matching URLs.

Also: Add unit tests. To simplify the testing, the parser has been
refactored to allow configuring it with custom tags. Still, the
default way to use it is as before, as a singleton.

https://phabricator.endlessm.com/T28687