libquest: Add app foreground utility functions
To detect when a t2 game starts or ends we can't use the app.is_running
because it uses the gtk dbus name and t2 games doesn't implement that
interface.

This patch modifies the method wait_for_app_in_foreground to accept a
new param `in_foreground`, so it could be used to wait for an app to be
in foreground or to wait for an app to be NOT in foreground, background
or closed.

There's also a new decorator with_app_in_foreground for methods in
quests to abort or run the final step when the quest app is finished.

https://phabricator.endlessm.com/T29008