clubhouse: Prevent switching to category view when quest is run
Previously when a narrative quest was set to play, the character
view switched back to the category view. This commit prevents
this behaviour.

Also, if the quest is run from other side like clicking on a link
quest://{quest_name}, we not only switch to the character view
but we set automatically the All category.

https://phabricator.endlessm.com/T28851