clubhouse: Update switch css on init
When init the clubhouse with the hack mode off the switch button text is
not synced and there's an inconsistence there showing the text ON and
the button in the OFF state.

This patch fixes the problem just updating the button css everytime the
lights changed, not only when the state is different.

https://phabricator.endlessm.com/T29076