tools: Simplify virtualenv usage
First, use the host's python3 to setup the virtualenv. Then, change
`run_in_python_venv` to simply call the program with the venv's `bin`
directory prepended to `PATH`. Part of the virtualenv setup is that the
shebang line on all the scripts is adjusted so it uses the `python` in
the virtualenv. That means the whole `activate`/`deactivate` dance is
unnecessary since the only critical thing it does is prepend `PATH` with
the venv's `bin` directory so that the appropriate `python` and tools
are found.

https://phabricator.endlessm.com/T29301