tools: Drop run_in_python_venv wrapper
Instead of adjusting `PATH` on the fly, just source the `activate` file
as if using the virtualenv interactively. This will adjust `PATH` and do
some other cleanup in the running shell. Then all the tools in the
virtualenv can be called normally without the wrapper since they will be
found in the virtualenv's `bin` directory normally.

https://phabricator.endlessm.com/T29301