tools: Add separate script to upgrade virtualenv
The logic in setup-venv was faulty and would always skip the upgrade
anyways. It could be fixed to always upgrade the packages, but that
would just slow everything down. Instead, add a separate script that
upgrades an existing virtualenv if a developer really wants to. In most
cases (and certainly in the CI) it's preferable to just wipe an existing
virtualenv directory and let it get recreated automatically.

https://phabricator.endlessm.com/T29301