quest, Sidetrack2: add detection for error fixed
To instruct the player step-by-step how to use
Flip-To-Hack add detection when the error has been
fixed and ask the player to flip back. Pulsate the
flip-button accordingly.

Do the same improvements to level 24.

Also make the pausing after unlocking greater so the
player has time to enjoy the toolbox.

https://phabricator.endlessm.com/T29253