libquest: Ignore insert_hyphens pango tag in narrative dialogues
Our runtime uses 1.42.3, but "insert_hyphens" had been added since
Pango 1.44 in 2e6f2eb8.

It only make sense to ignore "insert_hyphens" in narrative
dialogues because in the GNOME Shell we are using a newer version
of Pango, while using an older one for narrative dialogues.

This commit should be considered a workaround. Probably, the real
solution will be to update the runtime.

https://phabricator.endlessm.com/T29415