katamari: Update org.gnome.Sdk to 3.36
This change requires the update of gst-plugins-bad and soundtouch.
The gstreamer config options are not needed with meson because are auto
featured so the feauted will be built only if the dependency is there.

The gtksourceview is now included in the sdk, so we can remove that
dependency from the list.

https://phabricator.endlessm.com/T28433