system: Fallback to launch app outside EOS
This patch adds a fallback for the launch_app, if the app_launcher
doesn't work it because the dbus API isn't there. This patch adds a
custom script that's ran with flatpak-spawn --host to run the launch
script in the host.

This way we can launch any app, flatpaked or native app by id when we're
outside EOS.

https://phabricator.endlessm.com/T30458