clubhouse: Animate inapp notifications
This patch adds slide in from the left and movement animation when
moving the notification window to bottom/top.

The same notification window is used for all notifications, so the
window is not destroyed, just hidden. This makes the "next" flow a bit
smooth because the window is the same, so it's there always.

https://phabricator.endlessm.com/T28766