clubhouse: InAppNotify window not resizable
This change makes the window shrink to the size so when the dialog has a
long text the window will be bigger but when the text is shorter the
window will shrink to the minimum size.

This patch also fixes an issue with the notification window being
destroyed in some cases.

https://phabricator.endlessm.com/T28766