clubhouse: Add notificationEvent dbus API
This new method will allow us to send events from the shell extension to
the notification window. This is needed because windows are not shown on
overview mode so we need to show a clutter clone. To be able to
propagate click events to the real notification window we can use this
new method.

https://phabricator.endlessm.com/T28766