libquest: Add Tour extension waiters
This patch provides waiters functions to use in quests for hightlight
regions on the gnome-shell and wait for the user interaction, clicking
on the region or pressing skip.

The skip button will return False, so it can be used to determine if the
user click on the skip button or in the highlighted region. So it's
possible to detect if the skip button has been pressed like this:

  ret = self.wait_for_highlight_widget('clock')
  if not ret.future.result():
      # Skipped
      return self.skip_step

https://phabricator.endlessm.com/T30237