clubhouse: clean code notification access to private
This patch removes all the access to "private" date of the _msg
attribute in notification code.

Public properties has been added to the Message class to be able to
access to the move and close buttons. And some new methods has been
added to the InAppNotify class to set the internals on classmethod
constructors without calling directly to private attributes.

https://phabricator.endlessm.com/T28766