clubhouse: remove duplicate _init_style
This patch does the _init_style on the GtkApplication constructor so
it's not needed to do the same for the InAppNotify and we can remove the
duplicated code.

The call to _init_style should be done at construction time and not when
the window is created because it's possible to run a quest without the
clubhouse window.

https://phabricator.endlessm.com/T28766