libquest: Use new tag 'require:network' in quest
This new tag will be useful to differentiate the quests on the clubhouse
quest list when there's no connection.

This also moves the check for connection to the Quest class adding a new
method `step_no_connection` that can be overriden to run a different
step when there's no connection.

This new method is called instead of `step_begin` when there's no
connection and the quest requires network, and by default shows a
message and abort.

https://phabricator.endlessm.com/T30767