clubhouse: Refresh network status on every page change
This change should fix the problem with connectivity not in sync with
the real network status on EOS 3.8.

This patch adds a call to the NetworkManager DBus API to refresh the
network connectivity status every time the user changes the page in the
clubhouse.

This will update the cached value, so if you go to "Art", then turn off
the connection, go back and open "Art" again, you'll see the need
connection icon on quests.

https://phabricator.endlessm.com/T30758