tools: Add <a> tag to to hyperlinks
Fix links in messages for the cloned window (as opposed to messages in shell).

The shell was parsing the hyperlinks itself, with
`Util.findUrls`. When messages were in the shell, we only needed to add
markup for styling:

- Underline
- Foreground color
- Avoid inserting hyphen characters

For the cloned window, we do need to add the <a href="..." /> markup
in order to make links work.

https://phabricator.endlessm.com/T30809