libquest: Add show_highlight family method
The current highlight method for the quests are waiters, that require
the user to click on the highlighted region for the quest to continue.

This patch adds a set of new methods to be able to show a highlighted
region without a waiter, with a similar syntax, so the quest can
continue and show dialogs on top of that.

https://phabricator.endlessm.com/T30843