libquest: Add auto-install to wait_for_app_install
This patch adds two new optional parameters to the wait_for_app_install
method, confirm and repo.

Setting the confirm to False will launch the installation directly
without the need to click on Download on the App center.

The repo parameter should be used to select the flatpak remote
repository. By default is the flathub, but for other apps could be
eos-apps.

This will show the App center with the desired app and will start the
download inmediatly:

    self.wait_for_app_install(self.app, confirm=False, repo='eos-apps')

https://phabricator.endlessm.com/T30845