clubhouse: Fix in-app notification window placement
This patch just call to the _place method after the revealer slide-in
animation is completed, to place correctly the dialog in the final
position.

The closed signal is also moved to the end of the _close_message method,
to ensure that the dialog is hidden when this signal is emitted, so the
next dialog will slide-in. This is visible when you close a quest, the
daemon dialog will slide-in now.

https://phabricator.endlessm.com/T30855