clubhouse: Use hack extension notifications by default
The in app notification is not working correctly on wayland. There's no
easy way to fix that right now, so I think the best thing to do right
now to keep notifications working correctly is to just go back and use
the shell extension notifications.

This patch introduces a flag variable so we can enable or disable inapp
notifications, so the code is still there, but right now is just
disabled.

https://phabricator.endlessm.com/T30889