desktop: Don't override shell extension
If the old extension is installed we don't need to install the new one.
The old extension only exists on EOS3.8 so there's no need to install
the new extension here, the 3.8 version will work with the clubhouse.
Installing the new extension will cause some conflicts and break the
desktop, so it's better to keep the system one running.

https://phabricator.endlessm.com/T30957