metrics: Replace eosmetrics with matomo
This patch replaces the eosmetrics with a matomo server integration.

This patch also includes a queue for events, so we can store events if
there's no connection or if the matomo server is down or something like
that. This queue will be dequeued once an event is sent correctly.

To anonimize the data, a random unique user id is generated. This id is
stored on filesystem so we can use the same one everytime and we can
identify the same user without getting any real information about the
user.

This patch adds two new configuration variables to the clubhouse
module in the katamari tool:

 * matomo_host: The matomo server base url
 * matomo_site: The matomo site id

https://phabricator.endlessm.com/T30942