system App: Connect to the running change for non-gtk apps
In gtk apps we can connect a callback to the g_name_owner
property. But in non-gtk apps we have to do polling. We poll each 1
second to see if the name is owned in the dbus session through the
Desktop.app_is_running() method.

With this, libquest's Quest.wait_for_app_launch() and
Quest.ask_for_app_launch() work for non-gtk apps. For example:

    app = App('org.libreoffice.LibreOfficeIpc0', is_gtk=False)
    self.wait_for_app_launch(app)

Note that `is_gtk=False` has to be passed. Although it could be
automatically discovered using the org.freedesktop.DBus.Introspectable
interface, that provokes a gtk app to start, and app.is_running()
gives a false positive.

Also, now app.is_running() works for both gtk and non gtk apps.

https://phabricator.endlessm.com/T30904