system: Add delay to Desktop.focus_app
This is a workaround to fix the problem of focusing a desktop app the
first time.

There's a bug in gnome-shell and the first time we call the focus_app,
the page is not set correctly, so we need to add a delay to ensure that
the app grid is drawn before we call the focus_app.
https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/3388

https://phabricator.endlessm.com/T31083