metrics: Add Operating System as custom vars
The correct custom var parameter to use is _cvar. The _cvar format
should be a json dict of numbers and a dict with [key, value], something
like:

 {
  "1":["OS","iphone 5.0"],
  "2":["Matomo Mobile Version","1.6.2"],
  "3":["Locale","en::en"],
  "4":["Num Accounts","2"]
 }

You can see this data in the doc example:
https://developer.matomo.org/api-reference/tracking-api#example-tracking-request

This patch adds this _cvar value correctly with two variables for the
operating system and operating system version.

There's also a new parameter, custom, to been able to add other custom
variables as a dict to any record function.

https://phabricator.endlessm.com/T30942