achievemets: Only show notification if user-initiated
The achievement-achieved signal is emited for every achievement on
clubhouse startup because we don't record the achivements, we calculate
on runtime with the current state.

We only want to show the notification when this signal is emited after a
quest completion. That's done already for metrics so we should do the
same for the notifications.

This patch just adds a new boolean parameter to the achievement-achieved
signal that's true when the event was user initiated. This new parameter
is used to show the notification dialog.

https://phabricator.endlessm.com/T31194