libquest: Fix asyncio task for python 3.9
Without the loop parameter there's an error when trying to wait for
futures:

RuntimeError: no running event loop

It looks like the `loop` parameter is deprectated and will be removed in
python 3.10, but I didn't found another way to solve this problem. If we
don't pass the loop parameter to the wait method the
`events.get_running_loop()` inside the method raises the runtime error.

https://docs.python.org/3.9/library/asyncio-task.html#waiting-primitives