clubhouse: Handle empty RealName
If a user account's RealName property is empty, we would previously
crash while trying to pick the initials out of it:

    Traceback (most recent call last):
      File "/app/lib/python3.9/site-packages/eosclubhouse/clubhouse.py", line 3210, in do_activate
      File "/app/lib/python3.9/site-packages/eosclubhouse/clubhouse.py", line 3374, in _ensure_window
      File "/app/lib/python3.9/site-packages/eosclubhouse/clubhouse.py", line 2296, in __init__
      File "/app/lib/python3.9/site-packages/eosclubhouse/clubhouse.py", line 2438, in update_user_info
    IndexError: list index out of range

We would also set the empty string as the label for the user.

Fall back to the UserName, which is guaranteed to be non-empty, if
RealName is empty or all-whitespace.

(I encountered this problem in a WSL2 container, but the same problem
can be reproduced on any system.)